requirements.txt
fastapi==0.66.0
PyGithub==1.55
psycopg2==2.9.1
python-dotenv==0.18.0
pydantic==1.8.2
starlette==0.15.0
